+++ public/locales/en/translation.json
@@ -55,5 +55,7 @@
   "registration_select_pin_field_subtitle": "PIN are better than confirmation emails, right?",
   "resend_pin_button_label": "Resend PIN",
   "next": "Next",
-  "resend_pin": "Resend PIN in {{timeLeft}} seconds"
+  "resend_pin": "Resend PIN in {{timeLeft}} seconds",
+  "error_empty_field_default_message": "Input the PIN we sent to your email",
+  "error_regex_field_default_message": "This value in input is invalid"
 }
