+++ public/locales/es/translation.json
@@ -53,5 +53,7 @@
   "registration_select_avatar_field_title_hint": "¡Por ahora solo tenemos unas pocas opciones, pero el futuro podras generarlo a tu gusto!",
   "resend_pin_button_label": "Reenviar PIN",
   "next": "Siguiente",
-  "resend_pin": "Reenviar PIN en {{timeLeft}} segundos"
+  "resend_pin": "Reenviar PIN en {{timeLeft}} segundos",
+  "error_empty_field_default_message": "Este campo no puede estar vacio",
+  "error_regex_field_default_message": "Este campo no cumple con el formato requerido"
 }
