+++ src/ui/components/Globals/LogoComponent.jsx
